move tests in another file
